Add Platform Services Extension for iOS/Mac
Add an extension that splits out platform functionality
for third party frameworks depending on the target.
This adresses current incompatibilities with MSAppCenter
and OneSignal.
